requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.5.9
dnspython==2.7.0
Flask==2.1.2
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
langchain==0.0.222
langchainplus-sdk==0.0.20
loguru==0.7.2
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-enum==1.5.1
multidict==6.1.0
mypy-extensions==1.0.0
numexpr==2.10.2
numpy==1.26.4
openai==0.27.2
openapi-schema-pydantic==1.2.4
packaging==24.2
pinecone-client==2.2.1
propcache==0.2.0
pydantic==1.10.19
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
tenacity==8.5.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.1.2
yarl==1.18.0
